Final Model/requirements.txt
Mesa==0.8.9
networkx==2.5
numpy==1.20.1
PyYAML==6.0
